feat: move Routes.jsx from _components to components
