+++ src/components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Routes from '../_components/Routes';
+import Routes from './Routes';
 import './App.scss';
 import NavBar from './NavBar/NavBar';
