+++ src/components/Routes.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
 
-import Home from '../components/Home/Home';
-import Login from '../components/Auth/Login';
-import ResetPassword from '../components/ForgotPassword/ResetPassword';
-import ForgotPassword from '../components/ForgotPassword/ForgotPassword';
-import ForgotPasswordMessage from '../components/ForgotPassword/ForgotPasswordMessage';
+import Home from './Home/Home';
+import Login from './Auth/Login';
+import ResetPassword from './ForgotPassword/ResetPassword';
+import ForgotPassword from './ForgotPassword/ForgotPassword';
+import ForgotPasswordMessage from './ForgotPassword/ForgotPasswordMessage';
 
 const Routes = () => (
   <div>
